requirements.txt
dj-database-url==3.1.0
Django==6.0.1
djangorestframework==3.16.1
django-extensions==4.1
django-orderable==6.1.2
gunicorn==24.1.1
psycopg2-binary==2.9.11
whitenoise==6.11.0
